refactor(search): shared entity-type heading helper with i18n (#327)
